Change controlled reduction to Costa's method
Many changes, but the jist is:

* changed ControlledReduction.jl to mimic
Costa's code. In particular, it reduces the image
of each basis vector one by one, sweeping down
from high pole order to low pole order.

* farmed out a little bit of Utils.jl into
PolynomialWithPole.jl, which is for reasioning about
elements of k[x_1,...,x_n]_f (the localization).
Perhaps in the future that will house a custom struct,
but for now it's just some methods.