Fix global variable and degree bug
n and d were not defined in a couple of places,
but this was covered up by some
non-commented-out test code.
Now those are defined.

Moreover, d was sometimes defined as the total degree
with respect to x0 (or x1 if you're 1-indexed, whatever),
rather than the total degree of the polynomial f.

I think this has been a source of bugs, it's just that our examples
don't cover such cases.

Added a test to check for this, but something about zeta recovery
fails, so I added it to failing.csv