Add parallelization for naive strategy
Adds a struct that contains all the necessary info
for controlled reduction, and then uses that
to parallelize it in the naive strategy.

Right now, the code allocates one struct per vector.
One could mess around with the loops to make it allocate
one per thread, but this is a TODO for now.

Another TODO is to integrate the struct more fully with
the rest of the code, including in the costachunks
strategy (some work) and the naive strategy (symbol pushing)