Improve allocations in Linear Algebra
Make the linear algebra code allocate a lot less,
by making use of the gen_exp_vec cache that we've been using
elsewhere, and modifying Utils so that the methods which
are used in the linear algebra code allocate less.

Also fixed a lot of minor errors and updated other methods
to make things work.