Fix cubic fourfold experiment
Add manual calls to the garbage collector
to keep the experiment for finding cubic
fourfolds from running out of memory.

Also add some scaffolding to do the experiment on the GPU.
This includes moving `float_entries` to Utils.jl.
Right now, there seems to be something that causes
it to not work, but I'm not exactly sure what, it seems to be in
GPUFiniteFieldMatrices.jl.